Try getting rid of use of pybigwig
