+++ tests/datadriven/test_sequence_annotations.py
@@ -15,7 +15,6 @@ import tests.paths as paths
 
 import ga4gh.schemas.protocol as protocol
 
-import unittest
 
 _datasetName = "ds"
